Extract pagarme_facade, now you can instantiate Pagarme
